+++ src/console.rs
@@ -53,7 +53,7 @@ pub fn notify_site_size(site: &Site) {
         "-> Creating {} pages ({} orphan), {} sections, and processing {} images",
         library.pages().len(),
         site.get_number_orphan_pages(),
-        library.sections().len() - 1, // -1 since we do not the index as a section
+        library.sections().len() - 1, // -1 since we do not count the index as a section there
         site.num_img_ops(),
     );
 }